move hash base worker into new file
